+++ braid_stats/src/prior/pg.rs
@@ -1,5 +1,5 @@
 use rand::Rng;
-use rv::dist::{Gamma, InvGamma, Poisson};
+use rv::dist::{Gamma, Poisson};
 use rv::traits::*;
 use serde::{Deserialize, Serialize};
 
@@ -68,28 +68,28 @@ impl UpdatePrior<u32, Poisson, PgHyper> for Gamma {
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct PgHyper {
-    pub pr_shape: InvGamma,
-    pub pr_rate: InvGamma,
+    pub pr_shape: Gamma,
+    pub pr_rate: Gamma,
 }
 
 impl Default for PgHyper {
     fn default() -> Self {
         PgHyper {
-            pr_shape: InvGamma::new(1.0, 1.0).unwrap(),
-            pr_rate: InvGamma::new(1.0, 1.0).unwrap(),
+            pr_shape: Gamma::new(1.0, 1.0).unwrap(),
+            pr_rate: Gamma::new(1.0, 1.0).unwrap(),
         }
     }
 }
 
 impl PgHyper {
-    pub fn new(pr_shape: InvGamma, pr_rate: InvGamma) -> Self {
+    pub fn new(pr_shape: Gamma, pr_rate: Gamma) -> Self {
         PgHyper { pr_shape, pr_rate }
     }
 
     pub fn geweke() -> Self {
         PgHyper {
-            pr_shape: InvGamma::new_unchecked(10.0, 10.0),
-            pr_rate: InvGamma::new_unchecked(10.0, 10.0),
+            pr_shape: Gamma::new_unchecked(10.0, 10.0),
+            pr_rate: Gamma::new_unchecked(10.0, 10.0),
         }
     }
 
@@ -100,21 +100,15 @@ impl PgHyper {
         let m = xsf.iter().sum::<f64>() / nf;
         let v = xsf.iter().map(|&x| (x - m).powi(2)).sum::<f64>() / nf;
 
-        let a_shape = (2.0 * (m.powi(4) / v + 2.0)).max(2.0);
-        let a_scale = v / (m * m * (a_shape - 1.0));
+        let a_shape = (m + 1.0) * (m * m + 3.0 * m + v + 2.0) / v;
+        let b_shape = (m * m + 3.0 * m + 2.0 * v + 2.0) / v;
 
-        let b_shape = (0.1 * m * m) / (v * v) + 2.0;
-        let b_scale = m * (b_shape - 1.0) / v;
-
-        // Priors chosen so that mean of rate is the mean of the data and that
-        // the variance of rate is variance of the data. That is, we want the
-        // prior parameters μ = α/β and v = α^2/β
+        // Priors chosen so that the rate distribution has the mean and variance
+        // of the data
         PgHyper {
             // input validation so we can get a panic if something goes wrong
-            // pr_shape: InvGamma::new(2.0, v.recip()).unwrap(),
-            // pr_rate: InvGamma::new(2.0, v / m).unwrap(),
-            pr_shape: InvGamma::new(a_shape, a_scale).unwrap(),
-            pr_rate: InvGamma::new(b_shape, b_scale).unwrap(),
+            pr_shape: Gamma::new(a_shape, 1.0).unwrap(),
+            pr_rate: Gamma::new(b_shape, 1.0).unwrap(),
         }
     }