+++ Changelog.md
@@ -21,6 +21,7 @@
 - cli `codebook` function much faster
 - cli `codebook` function no longer supports the experimenal `Labeler` type
 - cli `codebook` function has `--no-check` argument to skip sanity checking
+- Count type hyper prior is Gamma(a, 1) on shape and Gamma(b, 1) on rate
 
 ## 0.32.3
 - Add '--flat-columns' argument to `braid run` command so that an engine can be